+++ src/Main.java
@@ -4,6 +4,7 @@ public class Main {
         Car lada = new Car("Lada",
                 "Granta",
                 0,
+                160,
                 "желтый",
                 2015,
                 "Россия",
@@ -14,7 +15,9 @@ public class Main {
                 new Car.Key(false,false));
         Car audi = new Car("Audi",
                 "A8 50 L TDI quattro",
-                3, "черный",
+                3,
+                220,
+                "черный",
                 0,
                 "Германия",
                 "Автомат",
@@ -25,6 +28,7 @@ public class Main {
         Car bmw = new Car("BMW",
                 "Z8",
                 3,
+                50,
                 "",
                 2021,
                 "Германия",
@@ -36,6 +40,7 @@ public class Main {
         Car kia = new Car("Kia",
                 "   ",
                 2.4,
+                400,
                 "красный",
                 2018,
                 "Южная Корея",
@@ -47,6 +52,7 @@ public class Main {
         Car hyundai = new Car(null,
                 "Avante",
                 1.6,
+                280,
                 "оранжевый",
                 2016,
                 "Южная Корея",
